+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/pgdbproject/parser/StubDatabaseLoader.java
@@ -20,6 +20,7 @@ import java.util.List;
 
 import org.pgcodekeeper.core.database.api.loader.ILoader;
 import org.pgcodekeeper.core.database.api.schema.IDatabase;
+import org.pgcodekeeper.core.settings.DiffSettings;
 import org.pgcodekeeper.core.settings.ISettings;
 
 /**
@@ -72,5 +73,10 @@ public class StubDatabaseLoader implements ILoader {
     @Override
     public List<Object> getErrors() {
         return Collections.emptyList();
+    }
+
+    @Override
+    public DiffSettings getDiffSettings() {
+        throw new IllegalStateException();
     }
 }